requirements.txt
DateTime==5.5
google-pasta==0.2.0
GoogleNews==1.6.15
keras==3.4.1
keras-tuner==1.4.7
pandas-datareader==0.10.0
pipreqs==0.5.0
pmdarima==2.0.4
scikeras==0.13.0
ta==0.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
tiingo==0.15.6
torch==2.4.0
yfinance==0.2.41

